Refactor Streamlit app structure by removing unused files and consolidating theme configuration
