+++ .streamlit/config.toml
@@ -3,3 +3,6 @@ primaryColor="#5ab2e6"
 backgroundColor="#f5f5f5"
 secondaryBackgroundColor="#63d463"
 textColor="#0a0a0a"
+showSidebarNavigation=false
+maxUploadSize=10
+font="sans serif"
