+++ requirements.txt
@@ -9,3 +9,4 @@ pandas>=1.5.0
 numpy>=1.23.0
 scikit-learn==1.3.2
 joblib>=1.2.0
+streamlit_option_menu
